DeviceSetValueCommand: new class
The command that sets the value of a particular device.
If system receives this command, selected device manager
will set value on required device (identified by DeviceID)
with given module (identified by moduleID). The setting
has to be done in given time (timeout). The individual
states of the setting value to which the command can get
need to be reported on server.

These individual states of the settings have not been implemented yet.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>