ServerLastValueCommand: new class
Some devices enable to set value on one or more of its modules
and after switching gateway on it is necessary to set these values
to the same values as in database.

For example, after the wireless socket is switched on, it is necessary
to set its status on the basis of the value from server because
the consistent information is required.

Signed-off-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>